Extracted TetrisBlock Widget into a new file and moved shapes into models directory.
