+++ lib/tetriminoWidget.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 
-import 'components/shapes.dart';
+import 'models/shapes.dart';
 
 class TetrisGamify extends StatefulWidget {
   const TetrisGamify({super.key});
